+++ src/libretro/drivers/sensor/__init__.py
@@ -1,2 +1,2 @@
-from .generator import *
-from .interface import *
+from .driver import *
+from .iterable import *