+++ CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 > breaking changes may be introduced
 > at any time without warning.
 
+## [Unreleased]
+
+### Changed
+
+- **BREAKING:** Rename `SensorInterface` to `SensorDriver`.
+- **BREAKING:** Rename `GeneratorSensorInterface` to `IterableSensorDriver`
+  and some accompanying type aliases.
+
 ## [0.4.0] - 2024-10-23
 
 ### Changed